Revert "OAuth2 client: add integration tests with Authelia (#8618)"
This reverts commit 842f973f943d7906ca03a5a4a0c3c76402e69727.
