--- api/client/src/intTest/resources/org/projectnessie/client/auth/oauth2/Dockerfile-authelia-version
@@ -1,3 +1,0 @@
-# Dockerfile to provide the image name and tag to a test.
-# Version is managed by Renovate - do not edit.
-FROM authelia/authelia:4.38@sha256:19375b10024caeef4e0b119a6247beae84cbaa02c846cfd750e92dea910d4b6a
