--- api/client/src/intTest/resources/org/projectnessie/client/auth/oauth2/authelia-cert.pem
@@ -1,30 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFDTCCAvWgAwIBAgIUTX+OKVRKOuAk9sEqWJyEAPYZBQYwDQYJKoZIhvcNAQEL
-BQAwFjEUMBIGA1UEAwwLZXhhbXBsZS5jb20wHhcNMjQwNTI0MTUzODE1WhcNMjUw
-NTI0MTUzODE1WjAWMRQwEgYDVQQDDAtleGFtcGxlLmNvbTCCAiIwDQYJKoZIhvcN
-AQEBBQADggIPADCCAgoCggIBAMVBBOoKMWufucJR0yHNthGqYKUoWYF2in+WS91Q
-qL0ENNbaHpagKHMDhWktxY+MBKn2Wt5j9q1Y2N/HMjPpQgpTkoFrhYwXB1GviXH+
-2M7pH5hpB8qZNGZnlWWEpjcKqwH2BKHCu8RfJlw/Ut6i2TYbvUROauFCx0W55gEn
-SRUMQkerffHD7M0d47Cr0lgD7FAcUUW8dF84nnV+b7qBchxYju0XUQyoyilwIdKd
-2P64oeGc3fAUHdyDr4tgwdpAF98ht4aI3E0JUQdGNfYu1K3cQskOz5llGJ/ACGEv
-6B55juRvYepsG7h8ATKyqawZ6zh98mx95bOa4c24tZB14QFVaAPs90amYMnOIf8D
-0o0YP//rpbAmyi8l+dQDiadu8/ZITceBV+9reBIw3/NNhJLEBc6TO6Slti4YtgIL
-cbHc7mqjTS/fAChEdtiRtm1oaDH/0lZHbHIBThhgSuZz7g2nzSt7qY1xEy5SHGOA
-XfpOIy7F5qEF0rf5Pqo/t/qXc99hOgI83l8niz3r6Dog3Rs/SoaLoftKesC1nxNx
-TwARoygzcYS3caEMnlfdnfpIoHBSQmK+BHnwS+9kj/dAGbnV78pSt/+/jHFx2NRJ
-c1iF6LKJbvzXgWIlP63ZlQJWE2zRMNZ81aRaJP9uTkbbGNSqsyV8Y3srrQ5xOULA
-6s/1AgMBAAGjUzBRMB0GA1UdDgQWBBS1QYTGUhd4go9muoCvQVt9zuYRUTAfBgNV
-HSMEGDAWgBS1QYTGUhd4go9muoCvQVt9zuYRUTAPBgNVHRMBAf8EBTADAQH/MA0G
-CSqGSIb3DQEBCwUAA4ICAQCHqCGhl6FHzJeocZOH56c1z+Y74Je6lLYWt1lq9onW
-px31Mp+Kjq9ztOgXD+f12pW23h+s129OYPxHiT1gEmKN7YhccRScMKjnpsnpVv1a
-Ytba+oD5w0qD3XCJ6eIjBzWlXkD9pPyTndl8vcO3lt0veMd8QEIc5RSDuVDGbK7A
-y8JI+0e7WT229rgC3xlAUNjPWBp2w9/yQZYp7QXyIuj2dBCMsjwQFPB9AOzMwhBK
-GT+he1fDwd43Rfxl44W5rzTHXyI0JXUxzApgoUOacM7+iLLHtACbH8mHj6ztv8MG
-WhaFEy1cvOOVMhldWJnQMGU1pYl1zCWRZ1vHy26Cy1EXey/YZuNb4iog2FA6lxKM
-k/6Pv0dprCgh6pWzRFZfzOEqO7JY9m2t3BwNQ/W57Dn5qQaPDVBdUIUSldYZoJLT
-5u0fECqBULBKK8fAsEAMOWYNAB2cwoB9Ijm9hhdYZFnly2ItZANmEZIUWfxbVEyH
-Mc9jCHtVpIe49Kb4ZLz/IpSY6O4O/lAJFlA0VSiTKUh3GVKZ+wzFAr3RevQgXps8
-tdkZSGYdKJLzg2uidEK2ODoqNHLyYqQ64rDG32LH0mvW3bGI3hQN3sl6Fcm87veW
-B6/yyKqUCfZCJ05wy4Aj4ra/ofWZ7tiBYrVfzg4/Qdy3QOlHvW8udwv3czyZg8YT
-Ww==
------END CERTIFICATE-----
