--- api/client/src/intTest/resources/org/projectnessie/client/auth/oauth2/authelia-key.pem
@@ -1,52 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIJQwIBADANBgkqhkiG9w0BAQEFAASCCS0wggkpAgEAAoICAQDFQQTqCjFrn7nC
-UdMhzbYRqmClKFmBdop/lkvdUKi9BDTW2h6WoChzA4VpLcWPjASp9lreY/atWNjf
-xzIz6UIKU5KBa4WMFwdRr4lx/tjO6R+YaQfKmTRmZ5VlhKY3CqsB9gShwrvEXyZc
-P1Leotk2G71ETmrhQsdFueYBJ0kVDEJHq33xw+zNHeOwq9JYA+xQHFFFvHRfOJ51
-fm+6gXIcWI7tF1EMqMopcCHSndj+uKHhnN3wFB3cg6+LYMHaQBffIbeGiNxNCVEH
-RjX2LtSt3ELJDs+ZZRifwAhhL+geeY7kb2HqbBu4fAEysqmsGes4ffJsfeWzmuHN
-uLWQdeEBVWgD7PdGpmDJziH/A9KNGD//66WwJsovJfnUA4mnbvP2SE3HgVfva3gS
-MN/zTYSSxAXOkzukpbYuGLYCC3Gx3O5qo00v3wAoRHbYkbZtaGgx/9JWR2xyAU4Y
-YErmc+4Np80re6mNcRMuUhxjgF36TiMuxeahBdK3+T6qP7f6l3PfYToCPN5fJ4s9
-6+g6IN0bP0qGi6H7SnrAtZ8TcU8AEaMoM3GEt3GhDJ5X3Z36SKBwUkJivgR58Evv
-ZI/3QBm51e/KUrf/v4xxcdjUSXNYheiyiW7814FiJT+t2ZUCVhNs0TDWfNWkWiT/
-bk5G2xjUqrMlfGN7K60OcTlCwOrP9QIDAQABAoICAEv+pccJUEr3iZis75vJZkje
-DEf1CK7RHcDUBbxpU+kmkrjn5fsADzVoeOCxyrhoKG/WwuGFP8c/QKbOVrAbQN87
-nWkRJK15Pc+hwmFR3vOwawTMCj1q0v5VEADBn0v7ZzDlOJmdrbuVMPE/UE0IDolD
-HKMRVyocESvy2AM/wG9EftCf6HDZnATWbB9nwy03rbabglBMr8TzylDX+nGvZ2g2
-UjJT1lz7No/KbK15NN3GAlAqoHLbPTn03g/joivUn53OnqdNWRqw6S/VQ/rpxfSJ
-tI1DkIW3pkn9sPzIeTQsfvnRU7b+GZv04n1GygnNHotwMpLL2lBc7nvUmtbiI2Iz
-/rNsZ7C2khgKizx06LQSMaRXAZ0NPcjM8gJrV9eB4kglunfC1ocISHu2evO1+9f6
-n++rKriKVm6QkEUZSXqb9yg+o5XL/6XdLkqMm7UEFQUHJRuirpuQ9D634wLc+gWD
-b/pPFJ6Glt+2iPvR85L31E7GJvBtL1wRSEgRNfZlifnd81KXec87TxyTFrBEVkOt
-hvVj61Nxt/J5thHZeK6r6j2W7HzcfFvw7ddu2C8mYQuh2fyQkNHQQKlLnlfcFiZV
-TKik8JVTPmRwHY8hwvDS+ReRcSpzygBNEY1jbYRpvZJf3JGLvsOq9rOnW2axDqjy
-sh327s3QXVeXgptvN1ntAoIBAQDiB+nbRdlI4Td4M72phWdTHiCRpfNLRiObdxuN
-/x9Va3F2g2djsw5xzW6+ky8d25dyt2PSV08JqKeY3bSI0j8UoRkBggF4DEGmi3rP
-QPxNDAli/PRg9cK9ByFPk5cOIiOB13GvE415uExN/3c/LcSWuDGIw9opssbZ8W3p
-hhC7/P3RfoQtfq2FuvEk6kSnAtv8Jgjh4qFEzrth/+HGuQ7ryHbbte1o7xUwlD01
-9pmuDySdG79Hu7XcEHS+4CY557TQPMv5Gus1HS4umfODqEWL/33wbc4aAA3H0PyF
-gYA6O2KTB0/xIVU4iUaeDv+eCXzphtCiF0WlegugsIUdaM8fAoIBAQDfaFcyN8cx
-uutRwCc2f7YE1BnOefkdeMUjcw/sw19ak8M+zIpNF3s+B1mDXgLI/J5rUFH2aW6n
-ANUTqc8FN7eXZ2eIyucwn5YQavAl0YnOwmY8Y8wZMmAlckWr96el01oJrVsmnwHe
-yLxtNPqpr+la/30/1fKWksOAFc/yFQBCaYOF2E1QT2SvltAYhv6FDP4kOOQzPtsx
-WMw8wLpqeKi4BOqoKjCJswRB5fHZOjNTzi4vEPo0s4Fx109pX5oWLzy94ncJ8FwD
-LtpRl6SAyVBrZdbBc6xkvmAqhAfa4BhLnn1jIS/qL8m5kW9I5AlqkgZGGuqDpFHX
-PSPrgxVxwgJrAoIBAQCfq6dv1q/f4/kjWKmqjyqmdKndObXcfve4YzN747Vmyjek
-2L64l0yk59JLnf2yweUCjLpfZH3Y/OMoAbX+0Ubuc78OaaBQBZh65h0jO2z/oHkh
-DEA2HmYq0laduE0FFuiuWFteqhlO2vpMPyMl5lbxgby8OF5tBnWikAJs7N/1bIqA
-6zbETLCXv6+JvFQzBroQLW4o48e4q7GDURdVInV/6mTMYIANmJcHcjCrZkC5zMi4
-k1lUey3E/3r5O67FSQFZ+Ri+KF3vb2WoVejIIJ7AfwOgTHFDvLX/aYp5ZToH6NuB
-sbCvTjS9c3wHO+Em6tzquum3F2mDs+ty4D+ovtlRAoIBAQDPnv56MxAyDU6oePV4
-VpPoRN1+4Cg91YEQgtxn4uH5IZ2v/RMCtjHY4ae0yGe6NdtcTBb/89fb6JAdaBB2
-/uLmGDlGDZtxur4Rn/BylRuHRW8R7IU/XWuxIVPSj6QEHjltCOjPKbLpQpCqxTO3
-Az7RTOMvsjty6A48OQk6J35IOhr7BpxDaKVf8b6PLNXY/MwrCZw6a831PFsCkvIs
-BNp3ukoB4sOdFPt96p8qjEYrOk/9eb+tvq0sCTdJ7Ikc2OWQaqjKWdCJPL+0WEOP
-VIgiZ1CAFGpAbhSdLnrGVMD/ddl9iNgowiStCpXrndGKcw7VlCoiEVHlwHJjaz1A
-rVblAoIBAHM7NB9ksC8MH6m6ckwVo7Cxifk7pxAGUdC45ulRR94JNyTU66S02Xy6
-Kt/uQYL4olpywh2GuEw50ZdaiWJ+Ie8hhDmEpY7wmv6voafGIAmhLhZDFE+W49eS
-Alk8qhQwxC2lnCLFpaR5/y7aIk67rbtpFGxluhEaAj8oxKWucc1ALAB5tgCJZ4Fg
-TetDk/6hXQFT7xDH43SeWYRcEht4XwSHYGCzH4le/hm2h+T85sPDdFoh4+xJmXbI
-xNR5pypZlZLzhfBVmhtuZSnP9El9kDWIAOEh59DngQpYBwMJjxp6ufRexVx126jD
-UUzdKLMVjXiqjrjwdPeZNzGbG48rTjo=
------END PRIVATE KEY-----
